# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/plans/like:
    put:
      tags:
      - rest-plan
      operationId: likePlan
      parameters:
      - name: abbrev
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
  /api/plans/likeC:
    put:
      tags:
      - rest-plan
      operationId: likePlanC
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/edited:
    put:
      tags:
      - rest-plan
      operationId: editPlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      - name: newDescription
        in: query
        required: true
        schema:
          type: string
      - name: abv
        in: query
        required: true
        schema:
          type: string
      - name: difficulty
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/dislike:
    put:
      tags:
      - rest-plan
      operationId: dislikePlan
      parameters:
      - name: abbrev
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
  /api/plans/dislikeC:
    put:
      tags:
      - rest-plan
      operationId: dislikePlanC
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/plans/new:
    post:
      tags:
      - rest-plan
      operationId: createPlan
      parameters:
      - name: category
        in: query
        required: true
        schema:
          type: string
      - name: planName
        in: query
        required: true
        schema:
          type: string
      - name: abv
        in: query
        required: true
        schema:
          type: string
      - name: description
        in: query
        required: true
        schema:
          type: string
      - name: difficulty
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not Found
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/done:
    post:
      tags:
      - rest-plan
      operationId: completePlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/plans:
    get:
      tags:
      - rest-plan
      operationId: getPlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/explore:
    get:
      tags:
      - rest-plan
      operationId: getPlansPage
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PagePlan_PlanDetails'
  /api/plans/completedPlans:
    get:
      tags:
      - rest-plan
      operationId: getAllCompletedPlan
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Completed_plan_CompletedPlanDetails'
  /api/plans/category:
    get:
      tags:
      - rest-plan
      operationId: getPlansbyCategoryName
      parameters:
      - name: category
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan_PlanDetails'
  /api/plans/:
    get:
      tags:
      - rest-plan
      operationId: getPlans
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/plans/completedPlan/removed:
    delete:
      tags:
      - rest-plan
      operationId: removeCompletedPlanbyUser
      parameters:
      - name: email
        in: query
        required: true
        schema:
          type: string
      - name: planName
        in: query
        required: true
        schema:
          type: string
      - name: date
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Completed_plan_CompletedPlanDetails'
...
components:
  schemas:
...
    Category_PlanDetails:
      type: object
      properties:
        name:
          type: string
        des:
          type: string
        color:
          type: string
        likedByUser:
          type: boolean
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan_PlanDetails'
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_PlanDetails'
        description:
          type: string
    Plan_PlanDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        likedUser:
          type: boolean
        category:
          $ref: '#/components/schemas/Category_PlanDetails'
    Tree_PlanDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_PlanDetails'
        category:
          $ref: '#/components/schemas/Category_PlanDetails'
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_PlanDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_PlanDetails'
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        password:
          type: string
          writeOnly: true
...
    PagePlan_PlanDetails:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        number:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Plan_PlanDetails'
        sort:
          $ref: '#/components/schemas/Sort_PlanDetails'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable_PlanDetails'
        empty:
          type: boolean
    Pageable_PlanDetails:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort_PlanDetails'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        paged:
          type: boolean
    Sort_PlanDetails:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    Category_CompletedPlanDetails:
      type: object
      properties:
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_CompletedPlanDetails'
        description:
          type: string
    Completed_plan_CompletedPlanDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_CompletedPlanDetails'
        email:
          type: string
        planName:
          type: string
    Tree_CompletedPlanDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_CompletedPlanDetails'
        category:
          $ref: '#/components/schemas/Category_CompletedPlanDetails'
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_CompletedPlanDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_CompletedPlanDetails'
        email:
          type: string
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        password:
          type: string
          writeOnly: true
...